ptst_requirements.txt
azure-cognitiveservices-language-luis==0.2.0
numpy==1.22.4
pandas==1.4.2
pytest==7.1.2
requests==2.25.1